Fixed genetic algorithm.
There was an issue with my algorithm. It was getting the same fitness
for evolving and best. This was because I mutated the dna for evolving
but I never updated evolving before calculating the fitness. So the
fitness was being calculated on old dna.

Added `RAND_PERCENT()`, a macro for a random in [0, 1]. See branch
`cool_macros` for some cool macros.